updated raycasting and movement

+++ kana_v03/src/render/raycasting.c
@@ -6,7 +6,7 @@
 /*   By: avoronko <avoronko@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/07 10:51:22 by avoronko          #+#    #+#             */
-/*   Updated: 2024/05/27 18:00:21 by avoronko         ###   ########.fr       */
+/*   Updated: 2024/05/29 10:57:42 by avoronko         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,8 +21,8 @@ void	init_ray(t_game *game, int current_x)
 	game->ray.dir_y = game->player.dir_y + game->player.plane_y * camera_x;
 	game->ray.delta_dist_x = fabs(1 / game->ray.dir_x);
 	game->ray.delta_dist_y = fabs(1 / game->ray.dir_y);
-	game->ray.pos_x = game->player.pos_x;
-	game->ray.pos_y = game->player.pos_y;
+	game->ray.pos_x = (int)game->player.pos_x;
+	game->ray.pos_y = (int)game->player.pos_y;
 }
 
 void	set_steps(t_game *game)
